Break out appeals table into separate, connected component
